requirements.txt
streamlit>=1.37
pandas>=2.2
sqlalchemy>=2.0
matplotlib>=3.8
plotly>=5.22
openai>=1.40
python-dotenv>=1.0
psycopg2-binary>=2.9
